requirements.txt
psutil
requests
websockets==10.1
python-dotenv
